requirements.txt
google-genai>=0.6.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
pyaudio>=0.2.14
google-api-python-client>=2.0.0
requests>=2.25.0

